+++ src/main/java/org/ureca/pinggubackend/domain/recruit/dto/request/RecruitPostDto.java
@@ -1,4 +1,4 @@
-package org.ureca.pinggubackend.domain.recruit.dto;
+package org.ureca.pinggubackend.domain.recruit.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
